Highlight trees if has_actual_sensor, small improvements
