--- superset-frontend/plugins/superset-plugin-trees-map/src/colors.ts
@@ -1,18 +1,0 @@
-export const ColorCollection = {
-  good: {
-    description: 'Gut',
-    color: [112, 187, 137],
-  },
-  average: {
-    description: 'Mäßig',
-    color: [224, 215, 126],
-  },
-  critical: {
-    description: 'Kritisch',
-    color: [254, 172, 118],
-  },
-  unknown: {
-    description: 'Unbekannt',
-    color: [229, 231, 235],
-  },
-};
